Conflicts:
	src/main/client/main.js

+++ src/main/java/com/conveyal/datatools/manager/controllers/api/OrganizationController.java
@@ -1,8 +1,10 @@
 package com.conveyal.datatools.manager.controllers.api;
 
+import com.conveyal.datatools.common.utils.SparkUtils;
 import com.conveyal.datatools.manager.auth.Auth0UserProfile;
 import com.conveyal.datatools.manager.models.JsonViews;
 import com.conveyal.datatools.manager.models.Organization;
+import com.conveyal.datatools.manager.models.Project;
 import com.conveyal.datatools.manager.utils.json.JsonManager;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -12,10 +14,16 @@ import spark.Request;
 import spark.Response;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Date;
+import java.util.HashSet;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
+import static com.conveyal.datatools.common.utils.SparkUtils.formatJSON;
 import static spark.Spark.*;
 import static spark.Spark.get;
 
@@ -38,8 +46,14 @@ public class OrganizationController {
 
     public static Collection<Organization> getAllOrganizations (Request req, Response res) {
         Auth0UserProfile userProfile = req.attribute("user");
+        boolean isOrgAdmin = userProfile.canAdministerOrganization();
         if (userProfile.canAdministerApplication()) {
             return Organization.getAll();
+        } else if (isOrgAdmin) {
+            List<Organization> orgs = new ArrayList<>();
+            orgs.add(userProfile.getOrganization());
+            LOG.info("returning org {}", orgs);
+            return orgs;
         } else {
             halt(401, "Must be application admin to view organizations");
         }
@@ -85,16 +99,55 @@ public class OrganizationController {
                 org.usageTier = Organization.UsageTier.valueOf(entry.getValue().asText());
             } else if(entry.getKey().equals("active")) {
                 org.active = entry.getValue().asBoolean();
-//            } else if(entry.getKey().equals("extensions")) {
-//                org.extensions = entry.getValue().asBoolean();
-//            } else if(entry.getKey().equals("subscriptionBeginDate")) {
-//                org.subscriptionBeginDate = Date.from;
+            } else if(entry.getKey().equals("extensions")) {
+                JsonNode extensions = entry.getValue();
+                Set<Organization.Extension> newExtensions = new HashSet<>();
+                for (JsonNode extension : extensions) {
+                    newExtensions.add(Organization.Extension.valueOf(extension.asText()));
+                }
+                org.extensions = newExtensions;
+            } else if(entry.getKey().equals("projects")) {
+                JsonNode projects = entry.getValue();
+                Collection<Project> projectsToInsert = new ArrayList<>(projects.size());
+                Collection<Project> existingProjects = org.getProjects();
+
+                // set projects orgId for all valid projects in list
+                for (JsonNode project : projects) {
+                    if (!project.has("id")) {
+                        halt(400, "Project not supplied");
+                    }
+                    Project p = Project.get(project.get("id").asText());
+                    if (p == null) {
+                        halt(404, "Project not found");
+                    }
+                    Organization previousOrg = p.getOrganization();
+                    if (previousOrg != null && !previousOrg.id.equals(org.id)) {
+                        halt(400, SparkUtils.formatJSON(String.format("Project %s cannot be reassigned while belonging to org %s", p.id, previousOrg.id), 400));
+                    }
+                    projectsToInsert.add(p);
+                    p.organizationId = org.id;
+                    p.save();
+                }
+                // assign remaining previously assigned projects to null
+                existingProjects.removeAll(projectsToInsert);
+                for (Project p : existingProjects) {
+                    p.organizationId = null;
+                    p.save();
+                }
+            } else if(entry.getKey().equals("subscriptionBeginDate")) {
+                org.subscriptionBeginDate = new Date(entry.getValue().asLong());
+            }  else if(entry.getKey().equals("subscriptionEndDate")) {
+                org.subscriptionEndDate = new Date(entry.getValue().asLong());
             }
         }
     }
 
     public static Organization deleteOrganization (Request req, Response res) {
         Organization org = requestOrganizationById(req);
+
+        if (org.getProjects().size() > 0) {
+            halt(400, formatJSON("Cannot delete organization that is referenced by projects.", 400));
+        }
         org.delete();
         return org;
     }
